+++ src/main/java/kr/co/samplepcb/xpse/resource/SpEstimateResource.java
@@ -120,4 +120,11 @@ public class SpEstimateResource {
     public CCResult selectPartnerEstimateItems(@RequestBody List<SpEstimatePartnerSelectionDTO> selections) {
         return this.spEstimateService.selectPartnerEstimateItems(selections);
     }
+
+    @Operation(summary = "견적서 + 협력사 견적서 검색", description = "견적서 목록을 협력사 견적서 하위 리스트와 함께 페이징 조건으로 검색합니다")
+    @JwtAuth
+    @GetMapping("/_searchWithPartners")
+    public CCPagingResult<SpEstimateListDTO> searchWithEstimate(Pageable pageable, SpEstimateSearchParam searchParam) {
+        return this.spEstimateService.searchWithPartners(pageable, searchParam);
+    }
 }
